changed users to auth and added new routes (users/current, users/update, users/subscribe)
