+++ .env.example
@@ -4,3 +4,4 @@ META_PASSWORD=
 CLOUDINARY_NAME=
 CLOUDINARY_KEY=
 CLOUDINARY_SECRET=
+PORT=
