+++ controllers/index.js
@@ -1,9 +1,11 @@
 const auth = require("./auth");
 const drinks = require("./drinks");
 const filters = require("./filters");
+const users = require("./users");
 
 module.exports = {
   auth,
   drinks,
   filters,
+  users,
 };
